requirements.txt
scikit-allel~=1.3.7
numpy~=1.26.4
zarr~=2.17.0
pyDataverse~=0.3.1